Extract CrawlStopConfirmationDialog to a separate component
Refactor the SourceDialog by moving the CrawlStopConfirmationDialog to its own component file. Add support for different stop actions (close or stop crawling) with improved state management and flexibility.
